move helpers to str
