+++ tests/test_str.py
@@ -1,5 +1,33 @@
 from pytest import raises
-from sc_crawler.str import extract_last_number, snake_case
+from sc_crawler.str import extract_last_number, snake_case, space_after, wrap
+
+
+def test_wrap_default_text():
+    assert wrap() == ""
+
+
+def test_wrap_empty_text():
+    assert wrap("") == ""
+
+
+def test_wrap_with_text():
+    assert wrap("foo") == " foo "
+
+
+def test_wrap_with_whitespace():
+    assert wrap("   foobar") == "    foobar "
+
+
+def test_wrap_with_before_after():
+    assert wrap("foo", before="__", after="__") == "__foo__"
+
+
+def test_space_after_default_text():
+    assert space_after() == ""
+
+
+def test_space_after_text():
+    assert space_after("foo") == "foo "
 
 
 def test_extract_last_number():
